Remove tree.py, adjust swc.py accordingly
